feat(admin,api): 2381 - [PDT] identification des lignes fusionnees
