+++ admin/src/scenes/plan-transport/ligne-bus/import/index.tsx
@@ -1,33 +1,25 @@
-import React, { useState, useEffect } from "react";
-import { useSelector } from "react-redux";
+import React, { useState } from "react";
+
+import SelectCohort from "@/components/cohorts/SelectCohort";
 
 import { Title } from "../../components/commons";
-import Select from "../../components/Select";
 import Download from "./components/Download";
 import Import from "./components/Import";
 import NavBar from "./components/NavBar";
 import Resum from "./components/Resum";
-import { getCohortSelectOptions } from "@/services/cohort.service";
-import { add } from "date-fns";
+import { ImportStep, ImportSummaryResponse } from "./type";
 
 export default function Index(props) {
-  const cohorts = useSelector((state) => state.Cohorts);
-  const [cohortList, setCohortList] = useState([]);
   const cohort = new URLSearchParams(props.location.search).get("cohort");
   const addLigne = new URLSearchParams(props.location.search).get("add");
-  const [steps, setSteps] = useState([
+  const [steps, setSteps] = useState<ImportStep[]>([
     { id: "étape 1", name: "Le modèle vierge", status: "current" },
     { id: "étape 2", name: "Mon plan de transport", status: "upcoming" },
     { id: "étape 3", name: "Résumé", status: "upcoming" },
   ]);
-  const [summary, setSummary] = useState(null);
+  const [summary, setSummary] = useState<ImportSummaryResponse | null>(null);
 
-  useEffect(() => {
-    const cohortList = getCohortSelectOptions(cohorts);
-    setCohortList(cohortList);
-  }, []);
-
-  const nextStep = () => {
+  const handleNextStep = () => {
     let nextStep = false;
     const newSteps = steps.map((step) => {
       if (step.status === "current") {
@@ -41,21 +33,21 @@ export default function Index(props) {
     setSteps(newSteps);
   };
 
-  function fileVerified(data) {
+  function handleFileVerified(data) {
     setSummary(data);
-    nextStep();
+    handleNextStep();
   }
 
   return (
     <div className="flex w-full flex-col px-8 pb-8 ">
       <div className="flex items-center justify-between py-8">
         <Title>{addLigne ? "Importer des lignes supplémentaires" : "Importer mon plan de transport"}</Title>
-        <Select options={cohortList} value={cohort} disabled={true} />
+        <SelectCohort cohort={cohort} disabled={true} />
       </div>
       <div className="mt-4 flex w-full flex-col">
         <NavBar steps={steps} />
-        {steps[0].status === "current" && <Download nextStep={nextStep} addLigne={addLigne} cohort={cohort} />}
-        {steps[1].status === "current" && <Import cohort={cohort} onFileVerified={fileVerified} addLigne={addLigne} />}
+        {steps[0].status === "current" && <Download onNextStep={handleNextStep} addLigne={addLigne} cohort={cohort} />}
+        {steps[1].status === "current" && <Import cohort={cohort} onFileVerified={handleFileVerified} addLigne={addLigne} />}
         {steps[2].status === "current" && <Resum cohort={cohort} summary={summary} addLigne={addLigne} />}
       </div>
     </div>
